backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
